// src/types/types.ts
export type AddressDataTypes = {
  id: string;
  firstName: string;
  lastName: string;
  companyName: string;
  address: string;
  city: string;
  state: string;
  countryName: string;
  postcode: string;
  email: string;
  country: string;
  addressType: string;
  vatId: string;
  phone: string;
  stateName: string;
  defaultAddress: boolean;
  useForShipping: boolean;
};